+++ src/Data/Functor/Rep.hs
@@ -63,6 +63,11 @@ module Data.Functor.Rep
   , duplicateRepBy
   , extendRepBy
   , extractRepBy
+  -- ** Generics
+  , GRep
+  , gindex
+  , gtabulate
+  , WrappedRep(..)
   ) where
 
 import Control.Applicative
@@ -92,9 +97,9 @@ import qualified Data.Sequence as Seq
 import Data.Semigroup hiding (Product)
 import Data.Tagged
 import Data.Void
+import qualified GHC.Generics as Gen
 import GHC.Generics hiding (Rep)
 import Prelude hiding (lookup)
-import qualified GHC.Generics as Gen
 
 -- | A 'Functor' @f@ is 'Representable' if 'tabulate' and 'index' witness an isomorphism to @(->) x@.
 --
@@ -109,22 +114,116 @@ import qualified GHC.Generics as Gen
 -- @
 
 class Distributive f => Representable f where
+  -- | If no definition is provided, this will default to 'GRep'.
   type Rep f :: *
-  type Rep f = Rep (Gen.Rep1 f)
+  type Rep f = GRep f
 
   -- |
   -- @
   -- 'fmap' f . 'tabulate' ≡ 'tabulate' . 'fmap' f
   -- @
+  --
+  -- If no definition is provided, this will default to 'gtabulate'.
   tabulate :: (Rep f -> a) -> f a
-  default tabulate :: (Gen.Generic1 f, Rep (Gen.Rep1 f) ~ Rep f, Representable (Gen.Rep1 f))
+  default tabulate :: (Gen.Generic1 f, GRep f ~ Rep f, GTabulate (Gen.Rep1 f))
                    => (Rep f -> a) -> f a
-  tabulate = Gen.to1 . tabulate
+  tabulate = gtabulate
 
+  -- | If no definition is provided, this will default to 'gindex'.
   index    :: f a -> Rep f -> a
-  default index :: (Gen.Generic1 f, Rep (Gen.Rep1 f) ~ Rep f, Representable (Gen.Rep1 f))
+  default index :: (Gen.Generic1 f, GRep f ~ Rep f, GIndex (Gen.Rep1 f))
                 => f a -> Rep f -> a
-  index = index . Gen.from1
+  index = gindex
+
+-- | A default implementation of 'Rep' for a datatype that is an instance of
+-- 'Generic1'. This is usually composed of 'Either', tuples, unit tuples, and
+-- underlying 'Rep' values. For instance, if you have:
+--
+-- @
+-- data Foo a = MkFoo a (Bar a) (Baz (Quux a)) deriving ('Functor', 'Generic1')
+-- instance 'Representable' Foo
+-- @
+--
+-- Then you'll get:
+--
+-- @
+-- 'GRep' Foo = Either () (Either ('WrappedRep' Bar) ('WrappedRep' Baz, 'WrappedRep' Quux))
+-- @
+--
+-- (See the Haddocks for 'WrappedRep' for an explanation of its purpose.)
+type GRep f = GRep' (Gen.Rep1 f)
+
+-- | A default implementation of 'tabulate' in terms of 'GRep'.
+gtabulate :: (Gen.Generic1 f, GRep f ~ Rep f, GTabulate (Gen.Rep1 f))
+          => (Rep f -> a) -> f a
+gtabulate = Gen.to1 . gtabulate'
+
+-- | A default implementation of 'index' in terms of 'GRep'.
+gindex :: (Gen.Generic1 f, GRep f ~ Rep f, GIndex (Gen.Rep1 f))
+       => f a -> Rep f -> a
+gindex = gindex' . Gen.from1
+
+type family GRep' (f :: * -> *) :: *
+class GTabulate f where
+  gtabulate' :: (GRep' f -> a) -> f a
+class GIndex f where
+  gindex' :: f a -> GRep' f -> a
+
+type instance GRep' (f :*: g) = Either (GRep' f) (GRep' g)
+instance (GTabulate f, GTabulate g) => GTabulate (f :*: g) where
+  gtabulate' f = gtabulate' (f . Left) :*: gtabulate' (f . Right)
+instance (GIndex f, GIndex g) => GIndex (f :*: g) where
+  gindex' (a :*: _) (Left  i) = gindex' a i
+  gindex' (_ :*: b) (Right j) = gindex' b j
+
+type instance GRep' (f :.: g) = (WrappedRep f, GRep' g)
+instance (Representable f, GTabulate g) => GTabulate (f :.: g) where
+  gtabulate' = Comp1 . tabulate . fmap gtabulate' . flip fmap WrapRep . curry
+instance (Representable f, GIndex g) => GIndex (f :.: g) where
+  gindex' (Comp1 fg) (i, j) = gindex' (index fg (unwrapRep i)) j
+
+type instance GRep' Par1 = ()
+instance GTabulate Par1 where
+  gtabulate' f = Par1 (f ())
+instance GIndex Par1 where
+  gindex' (Par1 a) () = a
+
+type instance GRep' (Rec1 f) = WrappedRep f
+instance Representable f => GTabulate (Rec1 f) where
+  gtabulate' f = Rec1 $ tabulate $ f . WrapRep
+instance Representable f => GIndex (Rec1 f) where
+  gindex' (Rec1 f) = index f . unwrapRep
+
+type instance GRep' (M1 i c f) = GRep' f
+instance GTabulate f => GTabulate (M1 i c f) where
+  gtabulate' = M1 . gtabulate'
+instance GIndex f => GIndex (M1 i c f) where
+  gindex' (M1 f) = gindex' f
+
+-- | On the surface, 'WrappedRec' is a simple wrapper around 'Rep'. But it plays
+-- a very important role: it prevents generic 'Distributive' instances for
+-- recursive types from sending the typechecker into an infinite loop. Consider
+-- the following datatype:
+--
+-- @
+-- data Stream a = a :< Stream a deriving ('Functor', 'Generic1')
+-- instance 'Representable' Stream
+-- @
+--
+-- With 'WrappedRep', we have its 'Rep' being:
+--
+-- @
+-- 'Rep' Stream = 'Either' () ('WrappedRep' Stream)
+-- @
+--
+-- If 'WrappedRep' didn't exist, it would be:
+--
+-- @
+-- 'Rep' Stream = Either () (Either () (Either () ...))
+-- @
+--
+-- An infinite type! 'WrappedRep' breaks the potentially infinite loop.
+newtype WrappedRep f = WrapRep { unwrapRep :: Rep f }
 
 {-# RULES
 "tabulate/index" forall t. tabulate (index t) = t #-}
@@ -311,12 +410,12 @@ instance Representable Par1 where
 
 instance Representable f => Representable (Rec1 f) where
   type Rep (Rec1 f) = Rep f
-  index (Rec1 f) i = index f i
+  index (Rec1 f) = index f
   tabulate = Rec1 . tabulate
 
 instance Representable f => Representable (M1 i c f) where
   type Rep (M1 i c f) = Rep f
-  index (M1 f) i = index f i
+  index (M1 f) = index f
   tabulate = M1 . tabulate
 
 newtype Co f a = Co { unCo :: f a } deriving Functor
@@ -364,24 +463,3 @@ liftR2 f fa fb = tabulate $ \i -> f (index fa i) (index fb i)
 
 liftR3 :: Representable f => (a -> b -> c -> d) -> f a -> f b -> f c -> f d
 liftR3 f fa fb fc = tabulate $ \i -> f (index fa i) (index fb i) (index fc i)
-
-instance (Representable a, Representable b, Distributive (a Gen.:*: b)) => Representable (a Gen.:*: b) where
-  type Rep (a Gen.:*: b) = Either (Rep a) (Rep b)
-  tabulate f = tabulate (f . Left) Gen.:*: tabulate (f . Right)
-  index (a Gen.:*: _) (Left i)  = index a i
-  index (_ Gen.:*: b) (Right i) = index b i
-
-instance Distributive Gen.Par1 => Representable Gen.Par1 where
-  type Rep Gen.Par1 = ()
-  tabulate f = Gen.Par1 $ f ()
-  index (Gen.Par1 x) () = x
-
-instance (Representable f, Distributive (Gen.Rec1 f)) => Representable (Gen.Rec1 f) where
-  type Rep (Gen.Rec1 f) = Rep f
-  tabulate f = Gen.Rec1 $ tabulate f
-  index (Gen.Rec1 x) i = index x i
-
-instance (Representable f, Distributive (Gen.M1 i c f)) => Representable (Gen.M1 i c f) where
-  type Rep (Gen.M1 i c f) = Rep f
-  tabulate f = Gen.M1 $ tabulate f
-  index = index